requirements.txt
asgiref==3.2.7
distro==1.5.0
Django==3.0.7
numpy==1.18.5
pandas==1.0.4
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
tabula-py==2.1.1

xhtml2pdf~=0.2.4